refactor(tangle-dapp): Move `LiquidStakingItem` to shared ui lib
