+++ apps/tangle-dapp/data/liquidStaking/adapters/moonbeam.tsx
@@ -1,6 +1,7 @@
 import { BN, BN_ZERO } from '@polkadot/util';
 import { createColumnHelper, SortingFnOption } from '@tanstack/react-table';
 import { LsProtocolId } from '@webb-tools/tangle-shared-ui/types/liquidStaking';
+import { LiquidStakingItem } from '@webb-tools/tangle-shared-ui/types/liquidStaking';
 import {
   Avatar,
   CheckBox,
@@ -15,7 +16,6 @@ import {
   LsParachainChainDef,
   LsToken,
 } from '../../../constants/liquidStaking/types';
-import { LiquidStakingItem } from '../../../types/liquidStaking';
 import { CrossChainTimeUnit } from '../../../utils/CrossChainTime';
 import formatBn from '../../../utils/formatBn';
 import { getApiPromise } from '../../../utils/polkadot';
