+++ apps/tangle-dapp/data/liquidStaking/adapters/phala.tsx
@@ -5,6 +5,7 @@ import {
   SortingFnOption,
 } from '@tanstack/react-table';
 import { LsProtocolId } from '@webb-tools/tangle-shared-ui/types/liquidStaking';
+import { LiquidStakingItem } from '@webb-tools/tangle-shared-ui/types/liquidStaking';
 import {
   Avatar,
   Chip,
@@ -18,7 +19,6 @@ import {
   LsParachainChainDef,
   LsToken,
 } from '../../../constants/liquidStaking/types';
-import { LiquidStakingItem } from '../../../types/liquidStaking';
 import { CrossChainTimeUnit } from '../../../utils/CrossChainTime';
 import formatBn from '../../../utils/formatBn';
 import { FetchProtocolEntitiesFn, GetTableColumnsFn } from '../adapter';
